+++ CHANGELOG.md
@@ -2,6 +2,31 @@
 
 All notable changes to this project will be documented in this file.
 
+# [2.0.0-beta.19] - 25-12-2019
+
+### Added
+
+- Table renderEmptyTable boolean prop
+
+### Fixes
+
+- Table header section did not have border when headerSection was omitted
+- Table headerSection option accepts null as a value
+
+# [2.0.0-beta.18] - 20-12-2019
+
+- Mistakenly omitted
+
+# [2.0.0-beta.17] - 17-12-2019
+
+### Fixes
+
+- InPlayerIcon name in Accordion
+
+### Added
+
+- Table header section
+
 # [2.0.0-beta.16] - 13-12-2019
 
 ### Change